🚚 Move some fee code around into more sensible places
